auth/requirements.txt
psycopg2-binary==2.9.3
Flask==2.1.2
gunicorn==20.1.0
gevent==21.12.0
Flask-Limiter[redis]==2.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Security==3.0.0
email-validator==1.2.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
flask-restx==0.5.1
Flask-JWT-Extended==4.4.1
redis==4.3.3
requests==2.28.1
tenacity==8.0.1
user-agents==2.2.0
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-exporter-jaeger==1.11.1
werkzeug==2.1.2